refactor: separate reward memory evaluation fixtures
